mv serde test helper to new test_utils crate
addresses review comment https://github.com/NomicFoundation/hardhat/pull/3998#discussion_r1235612107
