+++ crates/rethnet_rpc_server/Cargo.toml
@@ -16,3 +16,4 @@ tokio = { version = "1.21.2", default-features = false, features = ["macros"] }
 
 [dev-dependencies]
 bytes = { version = "1.2.1", default-features = false }
+rethnet_test_utils = { version = "0.1.0-dev", path = "../rethnet_test_utils" }
