+++ crates/rethnet_rpc_server/src/hardhat_methods.rs
@@ -126,34 +126,9 @@ pub enum HardhatMethodInvocation {
 mod tests {
     use bytes::Bytes;
 
+    use rethnet_test_utils::help_test_method_invocation_serde;
+
     use super::*;
-
-    fn help_test_method_invocation_serde(call: HardhatMethodInvocation) {
-        let json = serde_json::json!(call).to_string();
-
-        #[derive(Debug, serde::Deserialize)]
-        struct MethodInvocationStructWithUntypedParams {
-            #[allow(dead_code)]
-            method: String,
-            #[allow(dead_code)]
-            params: Vec<serde_json::Value>,
-        }
-        #[derive(Debug, serde::Deserialize)]
-        #[serde(untagged)]
-        enum MethodInvocationEnumWithUntypedParams {
-            Eth(MethodInvocationStructWithUntypedParams),
-            Hardhat(MethodInvocationStructWithUntypedParams),
-        }
-        serde_json::from_str::<MethodInvocationEnumWithUntypedParams>(&json).unwrap_or_else(|_| {
-            panic!(
-                "should have successfully deserialized, with params as a Vec<String>, json {json}"
-            )
-        });
-
-        let call_decoded: HardhatMethodInvocation = serde_json::from_str(&json)
-            .unwrap_or_else(|_| panic!("should have successfully deserialized json {json}"));
-        assert_eq!(call, call_decoded);
-    }
 
     #[test]
     fn test_serde_hardhat_add_compilation_result() {
