+++ crates/rethnet_test_utils/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rethnet_test_utils"
+version = "0.1.0-dev"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0.147", default-features = false, features = ["derive"] }
+serde_json = { version = "1.0.89" }
